feat: add bcrypt
