+++ .env
@@ -10,3 +10,4 @@ POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
 RESTART_POLYCY=always
 VERY_SECRET_KEY=secret
+SALT_ROUNDS=10
