+++ errLogs.txt
@@ -258,3 +258,95 @@ Catched unhandledRejection: Cannot render headers after they are sent to the cli
 Catched unhandledRejection: Cannot render headers after they are sent to the client
 
 
+2021-06-26T21:32:10.413Z
+Catched unhandledRejection: null value in column "name" of relation "user" violates not-null constraint
+
+
+2021-06-26T21:35:23.733Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T21:36:13.364Z
+Catched unhandledRejection: relation "user" does not exist
+
+
+2021-06-26T21:38:30.121Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T21:39:46.940Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T21:42:55.679Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T21:44:24.942Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T21:46:55.101Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T21:48:04.363Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:06:49.576Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:07:32.985Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:08:34.021Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:10:13.726Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:10:45.220Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:10:45.358Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:10:45.481Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:10:45.605Z
+Catched unhandledRejection: Cannot render headers after they are sent to the client
+
+
+2021-06-26T22:10:45.826Z
+Catched unhandledRejection: Cannot render headers after they are sent to the client
+
+
+2021-06-26T22:14:52.256Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:14:52.388Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:14:52.491Z
+Catched unhandledRejection: Cannot set headers after they are sent to the client
+
+
+2021-06-26T22:14:52.615Z
+Catched unhandledRejection: Cannot render headers after they are sent to the client
+
+
+2021-06-26T22:14:52.814Z
+Catched unhandledRejection: Cannot render headers after they are sent to the client
+
+
