+++ package.json
@@ -48,6 +48,7 @@
   },
   "dependencies": {
     "@types/uuid": "^8.3.0",
+    "bcrypt": "^5.0.1",
     "cors": "^2.8.5",
     "cross-env": "7.0.3",
     "dotenv": "^9.0.0",
@@ -64,6 +65,7 @@
     "yamljs": "0.3.0"
   },
   "devDependencies": {
+    "@types/bcrypt": "^5.0.0",
     "@types/express": "^4.17.12",
     "@types/express-jwt": "^6.0.1",
     "@types/jsonwebtoken": "^8.5.2",
