+++ src/resources/login/login.router.ts
@@ -2,26 +2,14 @@ import { Router } from 'express';
 import { getRepository } from 'typeorm';
 import jwt from 'jsonwebtoken';
 import dotenv from 'dotenv';
+import bcrypt from 'bcrypt';
 import { User } from '../../entites/user';
+import { createAdmin } from './login.service';
 
 dotenv.config();
 
 const { VERY_SECRET_KEY } = process.env;
 const router = Router();
-
-const createAdmin = async () => {
-  const userRepository = getRepository(User);
-  const user = await userRepository.findOne({ login: 'admin' });
-  if (!user) {
-    const admin = {
-      name: 'admin',
-      login: 'admin',
-      password: 'admin',
-    };
-    const newUser = await userRepository.create(admin);
-    await userRepository.save(newUser);
-  }
-};
 
 router.post('/', async (req, res) => {
   await createAdmin();
@@ -29,10 +17,11 @@ router.post('/', async (req, res) => {
   const { login, password } = req.body;
   const userRepository = getRepository(User);
   const user = await userRepository.findOne({ login });
-
   if (user) {
     const { id } = user;
-    if (user.password === password) {
+    const result = await bcrypt.compareSync(password, user.password);
+
+    if (result!) {
       const token = await jwt.sign({ id, login }, VERY_SECRET_KEY!);
       res.status(200).json({ token });
     } else {
@@ -40,6 +29,7 @@ router.post('/', async (req, res) => {
     }
     res.status(403).send('Unauthorized');
   }
+  res.status(403).send('Forbidden');
 });
 
 export default router;
